chm-rag/requirements.txt
beautifulsoup4>=4.12.0
html5lib>=1.1
lxml>=4.9.0
faiss-cpu>=1.7.4
openai>=1.0.0
tiktoken>=0.5.0
typer>=0.9.0
pyyaml>=6.0
rich>=13.0.0
numpy>=1.24.0
